Hacky fix for doc browser crashing
For some reason, this behavior changed in wxPython 3:
[using 3.0.2.0 msw (classic)]

  >>> wx.App()
  >>> frame = wx.Frame(None)
  >>> ctrl = wx.TextCtrl(frame)
  >>> ctrl.IsModified()
  False
  >>> ctrl = wx.TextCtrl(frame, style=wx.TE_RICH2)
  >>> ctrl.IsModified()
  True
  >>> # was False in wx 2.8

This seems like incorrect behavior to me - according to IsModified()'s
documentation, it's supposed to only return True if the user has
modified the text inside the control - which is clearly not the case,
since it's literally just been created.

Unfortunately, DocBrowser.DoSave() relies on this method to determine if
the user has changed text since that requires them to have entered a
file name as well. It then assumes that a valid file name exists, does a
sanity check and reports the error.

This commit just silences the error and aborts the save attempt, since
it's not really harmful to silently abort here. To properly fix this, we
would have to investigate _why_ this behavior changed.