Wrap all BitmapButton usages
This was way more painful and difficult than I imagined, mostly due to
this code bleeding into balt.DnDStatusBar, which _has_ to be the next
step in wrapping. It's full of _native_widget hacks now :(

Some steps that were needed in the process:
 - _AWidget had to learn how to change widget positions, since the
   status bar needs to do that to spread out the buttons. Alternatively,
   I could have reworked the status bar to use an HLayout, but that
   seemed like even more work...
 - Button had to learn how to handle right clicks, since the status bar
   buttons use that to create custom context menus.
 - Button had to learn how to hide borders and resize itself to exactly
   match its contents, since status bar buttons use this to create the
   illusion that they are 'clickable images'.
 - Finally, some utility classes wrapping the more complex abstractions
   had to be created (e.g. ClickableImage, ForwardButton, etc.).

On the bright side, the new ImageButton classes are much more powerful
than the old balt.bitmapButton stuff. They behave just like regular
buttons with images on them, meaning that they can even have text AND
images, in one of four different orientations.

Simplify ImageButton hierarchy

Gets rid of ImageButtonL, R, B and T in favor of just ImageButton. We
probably won't need the directions and even if we do, reintroducing them
as a parameter is far smarter and easier.

Drop hacky code in DnDStatusBar

This bug does not seem to happen with wx.Button - in fact, we now get a
double click since the hack fires a second event. I'm *extremely* happy
about being able to drop this, DnDStatusBar is an absolute nightmare, so
any simplification in it is *very* welcome.