ListMerger fixups and FIXMEs
Note in particular

+        self.srcMods = set(self.srcs) & p_file.loadSet # FIXME unused

Needs more work to move common code to common methods and cleanup/docs
Why initialization was completed later?

Note:

@@ -411,5 +411,3 @@ def finishPatch(self,patchFile, progress):
                             emptiesAdd(recordId)
-                if oldEmpties != empties:
-                    oldEmpties = empties.copy() # FIXME unused
-                    madeChanges = True
+                madeChanges |= oldEmpties != empties

Infernio:

Refactoring _AListsMerger a bit

Properly passes a defaultdict version of configChoices in instead of
hackily assigning it later and unifies the deflstMasters/delevMasters
variables as de_masters, eliminating some duplicate code.

Note this diff:
-        self.srcMods = set(self.srcs) & p_file.loadSet
+        self.srcs = set(self.srcs) & p_file.loadSet

That's almost certainly what was intended (it mirrors ListsMerger
perfectly), but the srcMods/srcs difference seems to have been missed
during the FO3/FNV merge.

Also dropped some commented out lines from the ListsMerger that were
still in the FidListsMerger and marked mastersScanned with a FIXME -
that was probably intended to avoid duplicate work for the de_masters
calculation, have to check if using it is feasible.

_AListsMerger couple docstring + fixup

Note that I dropped the itemgetter import, found that usage horribly
confusing, IMO the list comprehension is much clearer. Might be slower,
feel free to benchmark

Co-authored-by: Infernio <infernio@icloud.com>