Default Oblivion Patch now runs on 64bit
Tested with Oblivion and its DLCs and on the default settings for the
bashed patch. Further work is obviously needed but this serves as a good
staging point to understand what needs to change.

Added static typing to ctypes - performance will probably take a hit,
but did not profile either version. This lets us spot errors much easier
though, which is thoroughly worth it for this stage of the process.

First - there are tons of type errors in the ffi with CBash. Unsigned
cast as signed, mixing None and 0, etc. It seems(?) to run fine on 32bit
though, so not sure if it should be fixed.

Second - the problem lies squarely with those 'c_ulong' that Infernio
identified before. Some are in the functions he spotted, but there are
others spread about (in this run, the glaring ones are under those
'_ParentID' methods).

Re: c_void_p - it almost certainly works since we don't really care
about the contents themselves, but using those opaque structs like in
this commit should be prefered - even if we remove static typing on
arguments, using them for return types gives us a bit more safety, more
readability and shouldn't impact performance.