'All' Oblivion patchers now run on 64bit
Tested with Oblivion and its DLCs with 'Select All' patchers.

On the python only a few mistyped arguments needed to change while on
the C++ side a union was being initialized with the expectation that
int's size was the same as a pointer's. These changes are in the
'ganda-wip' branch in CBash and the compiled Debug dll is included in
this commit.