Rewrite ESL capability checking
Instead of going through our full machinery, we can devise a new, very
light class that *only* reads headers and discards everything else. In
the future, we may want to start using this class for UDR scanning etc.
as well (see the TODO).

This allows us to check *every* record type in SSE and FO4 - even ones
we don't have decoded for patching yet, since we only need to read the
headers to get the FormID, the rest is irrelevant.

Note that this commit may make a first startup *very* slow, since it
will have to check thousands of REFRs, NAVMs, etc., but once caching
kicks in it'll be just as fast as the current system.

Skip beth plugins for non-verbose mergeability checks

No need to waste time with these, none of them are mergeable or
ESL-capable. Will obviously have to be adjusted if a game shows up where
one of them *is* mergeable / ESL-capable. Note that this does *not*
apply to CC stuff - we currently don't have an easy way to retrieve info
about those, it's all internal stuff in _games_lo.py.