bain: Fix renamed plugins not notifying modInfos
While making delta patches to check some updated mods, I noticed that
the renamed plugins would not get the installer info on the Mods tab.
Cause was that the same method where I had to add BSA handling was using
the sources instead of the destinations, but since the file had been
renamed, the source was only valid for the temporary extraction
directory.

Also inlined some stuff since this is a very tight loop. Feel free to
benchmark, prove me wrong and drop them again :P