Update all bundled binaries RRR
1. 7zip to 19.00 64bit
In response to CVE-2018-10115.

Note that I used the 64bit version, since we're switching to 64bit
anyways and Igor has claimed it's ~10% faster:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/e164d46a/

2. LZMA SDK to 19.00
To bring it up to date with 7z above. Still the 32bit version, there
doesn't seem to be a 64bit one.

3. pngcrush to 64bit
Actually seems to be quite a bit faster, otherwise I wouldn't have
bothered.

Closes # 452. <--- RRR